fix check-state not updating
dataChanged signal must emit when reimplementing setData,
and if changed indexes not having same parent, behavior
is undefined.

Doc:
"If the items are of the same parent, the affected ones
are those between topLeft and bottomRight inclusive. If
the items do not have the same parent, the behavior is
undefined."

https://doc.qt.io/qt-5/qabstractitemmodel.html#dataChanged